// api_services/product/product.interface.ts
export interface GetSingleProductDto {
  id: string | number;
  business_id: string | number;
}
export interface GetSingleProductCommentsDto {
  id: string | number;
}

// Generated by https://quicktype.io

export interface SingleProductCommentDto {
  id: number;
  rate: number;
  title: null;
  message: string;
  reply_message: string;
  created_at: string;
  admin: Admin;
  user: User;
  status: Status;
}

export interface Admin {
  name: string;
  family: string;
}

export interface Status {
  id: number;
  step: number;
  title: string;
  description: null;
  color: string;
}

export interface User {
  id: number;
  first_name: null;
  last_name: null;
  mobile_number: string;
  email: null;
  birth_date: null;
  gender: null;
  credit: number;
  zip_codes: null;
  created_at: string;
  updated_at: string;
  email_verified_at: null;
  mobile_verified_at: string;
}

export interface SingleProductDto {
  id: number;
  unit_value: number;
  buy_price: number;
  in_cart?: number;
  sell_price: number;
  discount_amount: number;
  discount_percentage: string;
  view_count: number;
  unit: Unit;
  product: Product;
  business: Business;
  similar: SingleProductDto[];
  supplement: SingleProductDto[];
}

export interface Business {
  id: number;
  title: string;
  address: string;
  rate: string;
  location: Location;
  post_code: string;
  email: string;
  logo_location: string;
  has_deposit: boolean;
  send_price: number;
  send_type: string;
  coverage_post_codes: string[];
  commission_percentage: string;
  username: string;
  created_at: string;
}

export interface Location {
  latitude: string;
  longitude: string;
}

export interface Product {
  id: number;
  title: string;
  description: string;
  cover_location: CoverLocation;
  unit_value: number;
  code: string;
  barcode: string;
  rate: string;
  rate_count: number;
  buy_price: number;
  sell_price: number;
  discount_amount: number;
  discount_percentage: string;
  media_files: CoverLocation[];
  category: Brand;
  brand: Brand;
  depositItem: Brand;
  tags: Brand[];
  properties: Property[];
}

export interface Brand {
  id: number;
  title: string;
  image_location: null | string;
  created_at?: string;
  deposit_price?: number;
  type: "secondary" | "primary";
}

export interface CoverLocation {
  id: string;
  sort_number: null;
  file_location: string;
  file_mime_type: string;
}

export interface Property {
  id: number;
  title: string;
  value: string;
}

export interface Unit {
  id: number;
  title: string;
  key: string;
}
// ...
export type SendProductCommentDto = {
  rate: string | null | number;
  title: string | null;
  message: string | null;
};
export type AddToCartType = { business_product_id: string | number; count: string | number };

// Generated by https://quicktype.io

export interface PriceRangeData {
  min_price: number;
  max_price: number;
}

// Generated by https://quicktype.io

export interface AddToCartResponseDto {
  id: number;
  order_id: number;
  business_product_id: number;
  price: null;
  discount_amount: null;
  total_price: null;
  count: number;
  total_discount_amount: null;
  created_at: string;
  updated_at: string;
}

export type needBusinessProductId = { business_product_id: number | string };

// Generated by https://quicktype.io

export interface RateRangeproducts {
  total_count: number;
  rates: Rate[];
}

export interface Rate {
  rate: number;
  count: number;
}